Refactoring: new structure for client and server ctrl protocols
Removed the subfolders and packed 'em into single packages
